refactor: :truck: rename main-layout -> main-layouts
