+++ src/components/main-layouts/g-footer.tsx
